Add +/_ people

+++ client/app/admin/admin.controller.js
@@ -1,13 +1,13 @@
 'use strict';
 
 angular.module('timerDocFullstackApp')
-  .controller('AdminCtrl', function ($scope, $http, Auth, User, Doctor) {
+  .controller('AdminCtrl', function ($scope, $http, Auth, User, Doctor, Modal, socket) {
 
     $scope.doctor = {};
 
-    $scope.me = Auth.getCurrentUser();
     Doctor.get().success(function (data,status) {
-        $scope.doctors = data
+        $scope.doctors = data;
+        socket.syncUpdates('doctor', $scope.doctors);
     });
 
     $scope.addDoctor = function() {
@@ -27,12 +27,33 @@ angular.module('timerDocFullstackApp')
         });
     }
 
-    $scope.removeDoctor = function(doc) {
+    $scope.removeDoctor = Modal.confirm.delete(function(doc) {
         $http.delete('/api/doctors/'+doc._id);
-        angular.forEach($scope.doctors, function(u, i) {
+
+        /*angular.forEach($scope.doctors, function(u, i) {
             if (u === doc) {
                 $scope.doctors.splice(i, 1);
             }
+        });*/
+    });
+
+    $scope.addPatient = function(doctor) {
+        var count = doctor.nbPatient;
+        count = count+1;
+        $http.put('/api/doctors/'+doctor._id, {"nbPatient":count}).success(function(data, status) {
+            if(err) return console.log(err);
         });
     }
+
+    $scope.removePatient = function(doctor) {
+        var count = doctor.nbPatient;
+        count = count-1;
+        $http.put('/api/doctors/'+doctor._id, {"nbPatient":count}).success(function(data, status) {
+            if(err) return console.log(err);
+        });
+    }
+
+    $scope.$on('$destroy', function () {
+        socket.unsyncUpdates('doctor');
+    });
   });
